Added changes counter to memo_database.
Now memo_database is a struct containing a pointer to SQLite database and
a long integer representing a changes counter. The counter is incremented
each time an INSERT, UPDATE or DELETE query is sent to the database. It
will allow memo_word structs to automatically update themselves by
checking if the changes counter has been incremented since the last update.