Test of replies parsing functions has been passed.
The memo_check_reply function forks and executes a Perl script which parses
the body of an e-mail and extracts question-answer pairs. These are
subsequently read by the main process which checks if they match in the
database and updates positive or negative answers count.

Unfortunately the current implementation does not work before running make
install because parse-reply.pl cannot be found in $libdir/memo.